requirements.txt
airtable-python-wrapper==0.8.0
awsebcli==3.12.0
blessed==1.14.2
boto==2.28.0
botocore==1.7.36
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.9.3
MarkupSafe==0.23
pathspec==0.5.0
PyMySQL==0.6.3
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==0.9.8
tabulate==0.7.5
termcolor==1.1.0
todoist-python==7.0.17
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.9.6
WTForms==2.0.1
